gui: generate source code from .forms, not class files
The GUI Designer by default generates binary class files that
are added to the compiled variant of their classes. This adds
an unnecessary build dependency on intellij, as the GUI
designer is the only way to generate these files.

Instead of continuing to generate the binary class files,
generate the source files (which adds the uncompiled section)
for these swing forms. This removes the build dependency on
intellij.

However, when the forms are edited, they must be done through
intellij, and the contributor who does the edit must ensure
that the generate source files option is ticked, so that these
files remain in sync with their .forms.

Signed-off-by: Deven Bowers <bowerscd.source@gmail.com>